GDA-5177 - Updated hm model to have double rather than integer for value before and after in dark and flat fields. Parameter view contains many more fields.

+++ uk.ac.diamond.tomography.reconstruction.parameters.edit/src/uk/ac/diamond/tomography/reconstruction/parameters/hm/provider/DarkFieldTypeItemProvider.java
@@ -238,7 +238,7 @@ public class DarkFieldTypeItemProvider
 	 */
 	@Override
 	public String getText(Object object) {
-		BigDecimal labelValue = ((DarkFieldType)object).getValueBefore();
+		Double labelValue = ((DarkFieldType)object).getValueBefore();
 		String label = labelValue == null ? null : labelValue.toString();
 		return label == null || label.length() == 0 ?
 			getString("_UI_DarkFieldType_type") :
